separated essays and projects, need to find new place for buttons

+++ src/splash/ProjectExplorer.jsx
@@ -2,22 +2,15 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Table, Icon } from 'semantic-ui-react';
 import { connect } from 'react-redux';
-import { createTechFolioWindow, deleteFile, newTechFolioWindow } from '../techfolioeditor/TechFolioEditorWindow';
+import { createTechFolioWindow, deleteFile } from '../techfolioeditor/TechFolioEditorWindow';
 
 function handleClick(action, fileType, fileName) {
   if (action === 'edit') return createTechFolioWindow({ fileType, fileName });
-  if (action === 'delete') {
-    deleteFile(fileType, fileName);
-    return 0;
-  }
+  if (action === 'delete') return deleteFile(fileType, fileName);
   return -1;
 }
 
-function newClick(fileType) {
-  return newTechFolioWindow(fileType);
-}
-
-class FileExplorer extends React.Component {
+class ProjectExplorer extends React.Component {
   render() {
     const projectObjects = this.props.projects.map((project) => {
       return ({ key: `project-${project}`, fileName: project, fileType: 'projects' });
@@ -36,38 +29,30 @@ class FileExplorer extends React.Component {
       <Table celled unstackable striped>
         <Table.Header>
           <Table.Row>
-            <Table.HeaderCell colSpan={'2'} textAlign={'center'}>File Explorer</Table.HeaderCell>
+            <Table.HeaderCell textAlign={'center'}>Projects</Table.HeaderCell>
           </Table.Row>
         </Table.Header>
         <Table.Body>
-          {fileData.map(file => (
+          {projectObjects.map(file => (
             <Table.Row key={file.key}>
-              <Table.Cell>{file.fileName}</Table.Cell>
-              <Table.Cell textAlign={'right'}>
+              <Table.Cell>
                 <Icon link name="edit" onClick={() => handleClick('edit', file.fileType, file.fileName)} />
                 <Icon link name="delete" onClick={() => handleClick('delete', file.fileType, file.fileName)} />
+                {file.fileName}
               </Table.Cell>
             </Table.Row>))}
-          <Table.Row>
-            <Table.Cell selectable onClick={() => newClick({ fileType: 'projects' })}>
-              <a href={'#'}><b>Create a new project</b></a>
-            </Table.Cell>
-            <Table.Cell selectable onClick={() => newClick({ fileType: 'essays' })}>
-              <a href={'#'}><b>Create a new essay</b></a>
-            </Table.Cell>
-          </Table.Row>
         </Table.Body>
       </Table>
     );
   }
 }
 
-FileExplorer.defaultProps = {
+ProjectExplorer.defaultProps = {
   projects: [],
   essays: [],
 };
 
-FileExplorer.propTypes = {
+ProjectExplorer.propTypes = {
   projects: PropTypes.arrayOf(PropTypes.string),
   essays: PropTypes.arrayOf(PropTypes.string),
 };
@@ -79,4 +64,4 @@ function mapStateToProps(state) {
   };
 }
 
-export default connect(mapStateToProps)(FileExplorer);
+export default connect(mapStateToProps)(ProjectExplorer);
